+++ src/main/java/me/villagerunknown/immortalvillagers/Immortalvillagers.java
@@ -1,5 +1,8 @@
 package me.villagerunknown.immortalvillagers;
 
+import me.villagerunknown.immortalvillagers.feature.convertWitchToVillagerFeature;
+import me.villagerunknown.immortalvillagers.feature.nitwitEducationFeature;
+import me.villagerunknown.immortalvillagers.feature.resetVillagerTradesFeature;
 import me.villagerunknown.platform.Platform;
 import me.villagerunknown.immortalvillagers.feature.preventDamageToVillagersFeature;
 import me.villagerunknown.platform.PlatformMod;
@@ -32,5 +35,8 @@ public class Immortalvillagers implements ModInitializer {
 		
 		// # Activate Features
 		featureManager.addFeature( "preventDamageToVillagers", preventDamageToVillagersFeature::execute );
+		featureManager.addFeature( "convertWitchToVillager", convertWitchToVillagerFeature::execute );
+		featureManager.addFeature( "nitwitEducation", nitwitEducationFeature::execute );
+		featureManager.addFeature( "resetVillagerTrades", resetVillagerTradesFeature::execute );
 	}
 }